[FEATURE] move RoomParameters to pytta.classes.analysis and rename to RoomAnalysis
